+++ CHANGES.md
@@ -24,11 +24,14 @@ To be released.
 
 ### Bug fixes
 
+ -  Fixed a bug in `IStore.PruneOutdatedChains()`.  [[#3999]]
+
 ### Dependencies
 
 ### CLI tools
 
 [#3997]: https://github.com/planetarium/libplanet/pull/3997
+[#3999]: https://github.com/planetarium/libplanet/pull/3999
 
 
 Version 5.4.0